pyproject.toml
[project]
name = "doubao-image-mcp-server"
version = "1.0.0"
description = "Image generation MCP server based on Doubao API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.2.0",
    "volcengine-python-sdk[ark]>=1.0.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
]
authors = [{name = "suibin521", email = "your-email@example.com"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]

[project.urls]
Homepage = "https://github.com/suibin521/doubao-image-mcp-server"
Repository = "https://github.com/suibin521/doubao-image-mcp-server"
"Bug Tracker" = "https://github.com/suibin521/doubao-image-mcp-server/issues"
Documentation = "https://github.com/suibin521/doubao-image-mcp-server#readme"

[project.scripts]
doubao-image-mcp-server = "doubao_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["doubao_image_gen.py", "doubao_mcp_server.py"]

[tool.hatch.build.targets.sdist]
include = [
    "doubao_image_gen.py",
    "doubao_mcp_server.py",
    "README.md",
    "README_CN.md",
    "LICENSE",
    "images/",
]
exclude = [
    "log/",
    "*.log",
    "__pycache__/",
    "*.pyc",
    "*.pyo",
    "mcp_json.md",
]
